requirements.txt
fastapi[all]
uvicorn
pandas
faker
awsebcli
aws-cdk-lib>=2.0.0
constructs>=10.0.0
python-pptx
openai